refactor: split VSCode to be able to use only some parts
It allows to not have to call 'localize' ourselves
