+++ src/service-override/localization.ts
@@ -1,14 +1,14 @@
 import { IEditorOverrideServices } from 'vs/editor/standalone/browser/standaloneServices'
 import { SyncDescriptor } from 'vs/platform/instantiation/common/descriptors'
-import { ILanguagePackItem } from 'vs/platform/languagePacks/common/languagePacks'
+import { ILanguagePackItem, LanguagePackBaseService } from 'vs/platform/languagePacks/common/languagePacks'
 import { ILanguagePackService } from 'vs/platform/languagePacks/common/languagePacks.service'
 import { ILocaleService } from 'vs/workbench/services/localization/common/locale.service'
 import { IDialogService } from 'vs/platform/dialogs/common/dialogs.service'
 import { IHostService } from 'vs/workbench/services/host/browser/host.service'
 import { IProductService } from 'vs/platform/product/common/productService.service'
-import { localize, localizeWithPath } from 'vs/nls'
-import { Language, language } from 'vs/base/common/platform'
 import { URI } from 'vs/base/common/uri'
+import { AbstractLocaleService } from 'vs/workbench/services/localization/browser/localeService'
+import { IExtensionGalleryService } from 'vs/platform/extensionManagement/common/extensionManagement.service'
 import { getBuiltInExtensionTranslationsUris, setAvailableLocales } from '../l10n'
 import 'vs/workbench/contrib/localization/browser/localization.contribution'
 
@@ -23,85 +23,41 @@ export interface LocalizationOptions {
   availableLanguages: AvailableLanguage[]
 }
 
-class LocaleService implements ILocaleService {
-  _serviceBrand: undefined
-
+class LocaleService extends AbstractLocaleService {
   constructor (
     private options: LocalizationOptions,
-    @IDialogService private readonly dialogService: IDialogService,
-    @IHostService private readonly hostService: IHostService,
-    @IProductService private readonly productService: IProductService
+    @IDialogService dialogService: IDialogService,
+    @IHostService hostService: IHostService,
+    @IProductService productService: IProductService
   ) {
+    super(dialogService, hostService, productService)
   }
 
-  async setLocale (languagePackItem: ILanguagePackItem): Promise<void> {
-    const locale = languagePackItem.id
-
-    if (locale === Language.value() || (locale == null && Language.value() === navigator.language.toLowerCase())) {
-      return
-    }
-
+  override async storeLocale (locale: string | undefined): Promise<void> {
     if (locale == null) {
       await this.options.clearLocale()
     } else {
       await this.options.setLocale(locale)
     }
-
-    const restartDialog = await this.dialogService.confirm({
-      type: 'info',
-      message: localizeWithPath('vs/workbench/services/localization/browser/localeService', 'relaunchDisplayLanguageMessage', 'To change the display language, {0} needs to reload', this.productService.nameLong),
-      detail: localizeWithPath('vs/workbench/services/localization/browser/localeService', 'relaunchDisplayLanguageDetail', 'Press the reload button to refresh the page and set the display language to {0}.', languagePackItem.label),
-      primaryButton: localize({ key: 'reload', comment: ['&& denotes a mnemonic character'] }, '&&Reload')
-    })
-
-    if (restartDialog.confirmed) {
-      await this.hostService.restart()
-    }
   }
 
-  async clearLocalePreference (): Promise<void> {
+  override async clearLocale (): Promise<void> {
     await this.options.clearLocale()
-
-    const restartDialog = await this.dialogService.confirm({
-      type: 'info',
-      message: localizeWithPath('vs/workbench/services/localization/browser/localeService', 'clearDisplayLanguageMessage', 'To change the display language, {0} needs to reload', this.productService.nameLong),
-      detail: localizeWithPath('vs/workbench/services/localization/browser/localeService', 'clearDisplayLanguageDetail', "Press the reload button to refresh the page and use your browser's language."),
-      primaryButton: localize({ key: 'reload', comment: ['&& denotes a mnemonic character'] }, '&&Reload')
-    })
-
-    if (restartDialog.confirmed) {
-      await this.hostService.restart()
-    }
   }
 }
 
-class LanguagePackService implements ILanguagePackService {
-  _serviceBrand: undefined
-
+class LanguagePackService extends LanguagePackBaseService implements ILanguagePackService {
   constructor (
-    private options: LocalizationOptions
+    private options: LocalizationOptions,
+    @IExtensionGalleryService extensionGalleryService: IExtensionGalleryService
   ) {
+    super(extensionGalleryService)
     setAvailableLocales(new Set(options.availableLanguages.map(lang => lang.locale)))
   }
 
-  async getAvailableLanguages (): Promise<ILanguagePackItem[]> {
+  override async getAvailableLanguages (): Promise<ILanguagePackItem[]> {
     return this.options.availableLanguages.map(({ locale, languageName }) => {
-      const label = languageName ?? locale
-      let description: string | undefined
-      if (label !== locale) {
-        description = `(${locale})`
-      }
-
-      if (locale.toLowerCase() === language.toLowerCase()) {
-        description ??= ''
-        description += localizeWithPath('vs/platform/languagePacks/common/languagePacks', 'currentDisplayLanguage', ' (Current)')
-      }
-
-      return {
-        id: locale,
-        label,
-        description
-      }
+      return this.createQuickPickItem(locale, languageName)
     })
   }
